feat: :fire: massive refactor siuuuuu, better search bar and header
also reformatting the whole project soon!
